Added UserRepository, UserService/Impl(Retrieving external http service) (#2)
Added bet

+++ src/main/java/test/task/bets/model/User.java
@@ -13,5 +13,5 @@ public class User {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     Long id;
     String name;
-    Long valet;
+    Double valet;
 }
